Add surge protection (via replacement fee)
There could be the situation that we calculated a fee - but then a lot of transactions came in with higher fees - so ours does not get included and even swamped out the mempool rendering the faucet in a non -working state as there is a nonce gap.
This commit should fix this by replacing the fee after some time if needed.
Side effect: as the tx hash can change then we only present the tx hash to the user once it is included in a block
